chore: refactor image file structure
