+++ src/pages/careers/_model/_locations/_locations.tsx
@@ -27,14 +27,14 @@ import MalaysiaFlagIcon from 'images/svg/careers/flag_malaysia.svg'
 import ParaguayFlagIcon from 'images/svg/careers/flag_paraguay.svg'
 import UAEFlagIcon from 'images/svg/careers/flag_uae.svg'
 import MaltaFlagIcon from 'images/svg/careers/flag_malta.svg'
-import CyprusFlagIcon from 'images/svg/careers/flag_cyprus.svg'
-import CaymanFlagIcon from 'images/svg/careers/flag_cayman.svg'
+import CyprusFlagIcon from 'images/svg/careers/career_flag_cyprus.svg'
+import CaymanFlagIcon from 'images/svg/careers/career_flag_cayman.svg'
 import RwandaFlagIcon from 'images/svg/careers/flag_rwanda.svg'
-import BVIFlagIcon from 'images/svg/careers/flag_bvi.svg'
+import BVIFlagIcon from 'images/svg/careers/career_flag_bvi.svg'
 import UKFlagIcon from 'images/svg/careers/flag_uk.svg'
-import FranceFlagIcon from 'images/svg/careers/flag_france.svg'
+import FranceFlagIcon from 'images/svg/careers/career_flag_france.svg'
 import ChannelIslandsFlagIcon from 'images/svg/careers/flag._channel_islands.svg'
-import GermanyFlagIcon from 'images/svg/careers/flag_germany.svg'
+import GermanyFlagIcon from 'images/svg/careers/career_flag_germany.svg'
 import VanuatuFlagIcon from 'images/svg/careers/flag_vanuatu.svg'
 import JordanFlagIcon from 'images/svg/careers/flag_jordan.svg'
 import SingaporeFlagIcon from 'images/svg/careers/flag_singapore.svg'
