+++ src/pages/careers/_our-hiring-process.tsx
@@ -3,17 +3,17 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import { zoho_url } from 'common/constants'
 // SVG
-import Career1 from 'images/svg/careers/career-1.svg'
-import Career2 from 'images/svg/careers/career-2.svg'
-import Career3 from 'images/svg/careers/career-3.svg'
-import Career4 from 'images/svg/careers/career-4.svg'
-import DownwardsLeft from 'images/svg/careers/downwards-left.svg'
-import DownwardsRight from 'images/svg/careers/downwards-right.svg'
+import Career1 from 'images/svg/careers/home/home_1.svg'
+import Career2 from 'images/svg/careers/home/home_2.svg'
+import Career3 from 'images/svg/careers/home/home_3.svg'
+import Career4 from 'images/svg/careers/home/home_4.svg'
+import DownwardsLeft from 'images/svg/careers/career_downwards_left.svg'
+import DownwardsRight from 'images/svg/careers/career_downwards_right.svg'
 import { Container, SectionContainer, Flex } from 'components/containers'
 import { Text, Header, LinkText } from 'components/elements'
 import { LinkButton } from 'components/form'
 import device from 'themes/device'
-import TipIcon from 'images/svg/careers/career-tip.svg'
+import TipIcon from 'images/svg/careers/career_tip.svg'
 import useRegion from 'components/hooks/use-region'
 
 const StyledSectionContainer = styled(SectionContainer)`
